+++ C/main.c
@@ -25,8 +25,8 @@ int main(int argc,char *argv[ ]){
 
     int seed = atoi(argv[1]);
     double N = atof(argv[2]);
-    int redes = atoi(argv[3]);
-    int vacina = atoi(argv[4]);
+    uint16_t redes = atoi(argv[3]);
+    uint8_t vacina = atoi(argv[4]);
     double prob = atof(argv[5])/100;
     bool weight = atoi(argv[6]);
     generate_infect(N,prob,seed, redes,vacina,weight);
